+++ CHANGELOG.md
@@ -4,4 +4,6 @@ All notable changes to `AppUI` will be documented in this file.
 
 ## 1.0.0 - 202X-XX-XX
 
-- initial release
+- Layout components
+- Table components
+- Card panel